Extract null reading tests
